Moved shaders and opengl stuff into Window
