+++ pycraft/window.py
@@ -7,6 +7,7 @@ from pyglet.gl import *
 from pyglet.window import key, mouse
 
 from pycraft.objects import brick, grass, sand, stone
+from pycraft.shader import Shader
 from pycraft.util import sectorize, cube_vertices
 
 TICKS_PER_SEC = 60
@@ -21,6 +22,9 @@ class Window(pyglet.window.Window):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
+        self.init_gl()
+        self.shader = None
+        self.init_shader()
         self.set_world(None)
         self.set_player(None)
         # The crosshairs at the center of the screen.
@@ -35,6 +39,56 @@ class Window(pyglet.window.Window):
         # This call schedules the `update()` method to be called
         # TICKS_PER_SEC. This is the main game event loop.
         pyglet.clock.schedule_interval(self.update, 1.0 / TICKS_PER_SEC)
+
+    def init_gl(self):
+        """Basic OpenGL configuration."""
+        # Set the color of "clear", i.e. the sky, in rgba.
+        glClearColor(0.5, 0.69, 1.0, 1)
+        # Enable culling (not rendering) of back-facing facets -- facets that aren't
+        # visible to you.
+        glEnable(GL_CULL_FACE)
+        # Set the texture minification/magnification function to GL_NEAREST (nearest
+        # in Manhattan distance) to the specified texture coordinates. GL_NEAREST
+        # "is generally faster than GL_LINEAR, but it can produce textured images
+        # with sharper edges because the transition between texture elements is not
+        # as smooth."
+        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST)
+        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST)
+        self.init_gl_fog()
+
+    def init_gl_fog(self):
+        """Configure the OpenGL fog properties."""
+        # Enable fog. Fog "blends a fog color with each rasterized pixel fragment's
+        # post-texturing color."
+        glEnable(GL_FOG)
+        # Set the fog color.
+        glFogfv(GL_FOG_COLOR, (GLfloat * 4)(0.5, 0.69, 1.0, 1))
+        # Say we have no preference between rendering speed and quality.
+        glHint(GL_FOG_HINT, GL_DONT_CARE)
+        # Specify the equation used to compute the blending factor.
+        glFogi(GL_FOG_MODE, GL_LINEAR)
+        # How close and far away fog starts and ends. The closer the start and end,
+        # the denser the fog in the fog range.
+        glFogf(GL_FOG_START, 20.0)
+        glFogf(GL_FOG_END, 60.0)
+
+    def init_shader(self):
+        vertex_shader = ""
+        fragment_shader = ""
+
+        with open("pycraft/shaders/world.vert") as handle:
+            vertex_shader = handle.read()
+
+        with open("pycraft/shaders/world.frag") as handle:
+            fragment_shader = handle.read()
+
+        self.shader = Shader([vertex_shader], [fragment_shader])
+
+    def start_shader(self):
+        self.shader.bind()
+
+    def stop_shader(self):
+        self.shader.unbind()
 
     def set_world(self, world):
         self.world = world
@@ -153,9 +207,9 @@ class Window(pyglet.window.Window):
         self.clear()
         self.set_3d()
         glColor3d(1, 1, 1)
-        self.world.start_shader()
+        self.start_shader()
         self.world.batch.draw()
-        self.world.stop_shader()
+        self.stop_shader()
         self.draw_focused_block()
         self.set_2d()
         self.draw_label()
