+++ pycraft/world.py
@@ -9,7 +9,6 @@ from pyglet.window import mouse
 
 from pycraft.objects import brick, grass, sand, stone
 from pycraft.util import normalize, sectorize, cube_vertices, cube_shade
-from pycraft.shader import Shader
 
 simplex_noise2 = SimplexNoise(256).noise2
 
@@ -42,46 +41,10 @@ class World:
         self.sector = None
         # Mapping from sector to a list of positions inside that sector.
         self.sectors = {}
-
-        self.shader = None
         # Simple function queue implementation. The queue is populated with
         # _show_block() and _hide_block() calls
         self.queue = deque()
-        self.init_gl()
         self._initialize()
-        self.init_shader()
-
-    def init_gl(self):
-        """Basic OpenGL configuration."""
-        # Set the color of "clear", i.e. the sky, in rgba.
-        glClearColor(0.5, 0.69, 1.0, 1)
-        # Enable culling (not rendering) of back-facing facets -- facets that aren't
-        # visible to you.
-        glEnable(GL_CULL_FACE)
-        # Set the texture minification/magnification function to GL_NEAREST (nearest
-        # in Manhattan distance) to the specified texture coordinates. GL_NEAREST
-        # "is generally faster than GL_LINEAR, but it can produce textured images
-        # with sharper edges because the transition between texture elements is not
-        # as smooth."
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST)
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST)
-        self.init_gl_fog()
-
-    def init_gl_fog(self):
-        """Configure the OpenGL fog properties."""
-        # Enable fog. Fog "blends a fog color with each rasterized pixel fragment's
-        # post-texturing color."
-        glEnable(GL_FOG)
-        # Set the fog color.
-        glFogfv(GL_FOG_COLOR, (GLfloat * 4)(0.5, 0.69, 1.0, 1))
-        # Say we have no preference between rendering speed and quality.
-        glHint(GL_FOG_HINT, GL_DONT_CARE)
-        # Specify the equation used to compute the blending factor.
-        glFogi(GL_FOG_MODE, GL_LINEAR)
-        # How close and far away fog starts and ends. The closer the start and end,
-        # the denser the fog in the fog range.
-        glFogf(GL_FOG_START, 20.0)
-        glFogf(GL_FOG_END, 60.0)
 
     def _initialize(self):
         """Initialize the world by placing all the blocks."""
@@ -323,21 +286,3 @@ class World:
         """Process the entire queue with no breaks."""
         while self.queue:
             self._dequeue()
-
-    def init_shader(self):
-        vertex_shader = ""
-        fragment_shader = ""
-
-        with open("pycraft/shaders/world.vert") as handle:
-            vertex_shader = handle.read()
-
-        with open("pycraft/shaders/world.frag") as handle:
-            fragment_shader = handle.read()
-
-        self.shader = Shader([vertex_shader], [fragment_shader])
-
-    def start_shader(self):
-        self.shader.bind()
-
-    def stop_shader(self):
-        self.shader.unbind()
